requirements.txt
fastapi
uvicorn
python-dotenv
playwright
yagmail
jinja2
